Move controller logic to a helper
This will allow to unit test this code.
