+++ app/views/coronavirus_form/check_answers.html.erb
@@ -16,7 +16,7 @@
 } %>
 
 <%= render "govuk_publishing_components/components/summary_list", {
-  items: @items,
+  items: answer_items,
 } %>
 
 <%= render "govuk_publishing_components/components/heading", {
